Highlight municipality polygons containing OSINT events
Fetch muni_il_compressed.json, use turf point-in-polygon to find which
municipalities contain event locations, and light up only those polygons.
